Extra text
origin example had this text removed but didn't specify to delete any text. upon seeing red and green highlights in the example after I noticed that some text was deleted so decided to delete the same type of text to see the same type of results. consistency is key to be able to understand exactly what's going on here. but because of that unclear direction i had to create this additional commit.